Move GlobUtils to com.oracle.svm.util
