+++ substratevm/src/com.oracle.svm.util/src/com/oracle/svm/util/GlobUtils.java
@@ -23,11 +23,9 @@
  * questions.
  */
 
-package com.oracle.svm.core.jdk.resources.CompressedGlobTrie;
+package com.oracle.svm.util;
 
 import java.util.List;
-
-import com.oracle.svm.util.StringUtil;
 
 public class GlobUtils {
